ProductCatalogService Cleanup (#317)
